% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_ratio.correction.R
\name{batch_ratio.correction}
\alias{batch_ratio.correction}
\title{QC-based BEC algorithm: batch-ratio}
\usage{
batch_ratio.correction(data, method = c("ratio-A", "median", "mean"))
}
\arguments{
\item{data}{A dataframe.\cr
Execute \code{data(Dataset_I)} and \code{View(Dataset_I)} for formats.}

\item{method}{Default is \code{ratio-A}, which combines median values and mean values, and is unnecessary to be tuned for most of the time.}
}
\value{
A dataframe. The corrected data, whose format is the same as \code{data}.
}
\description{
Batch-ratio is used for inter-BEC, which has several versions. Here, we provide three versions and recommend to use \code{ratio-A}.
}
\note{
\itemize{
 \item{If \code{method = ratio-A} or \code{mean}, it can ensure the consistency among QC samples' mean vectors across different batches after correction.}
}
}
\examples{
data(Dataset_I)
data <- Dataset_I
data.batch_ratio <- batch_ratio.correction(data)
}
\references{
\itemize{
 \item{van der Kloet, F.; Bobeldijk, I.; Verheij, E.; Jellema, R. Analytical Error Reduction Using Single Point Calibration for Accurate and Precise Metabolomic Phenotyping. \emph{Journal of Proteome Research} \strong{2009}, 8, 5132-5141. DOI: 10.1021/pr900499r.}
 \item{Kamleh, M. A.; Ebbels, T. M. D.; Spagou, K.; Masson, P.; Want, E. J. Optimizing the Use of Quality Control Samples for Signal Drift Correction in Large-Scale Urine Metabolic Profiling Studies. \emph{Analytical Chemistry} \strong{2012}, 84 (6), 2670-2677. DOI: 10.1021/ac202733q.}
 \item{Wang, S.-Y.; Kuo, C.-H.; Tseng, Y. J. Batch Normalizer: A Fast Total Abundance Regression Calibration Method to Simultaneously Adjust Batch and Injection Order Effects in Liquid Chromatography/Time-of-Flight Mass Spectrometry-Based Metabolomics Data and Comparison with Current Calibration Methods. \emph{Analytical Chemistry} \strong{2012}, 85 (2), 1037-1046. DOI: 10.1021/ac302877x.}
}
}
\author{
Zhendong Guo (\email{guozhendong19@mails.ucas.ac.cn}).
}
