% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance_correction.R
\name{CoCo}
\alias{CoCo}
\title{QC-based BEC algorithm: Covariance Correction (CoCo)}
\usage{
CoCo(
  data,
  alpha_range = c(0, 1),
  lambda_range = c(0, 10),
  search_iterations = 500,
  random_seed = 123,
  continue = FALSE,
  sig_level.alpha = 0.05,
  simul_method = c(NULL, "cauchy", "fisher", "HN"),
  p_value.adjust = c("bonferroni", "holm", "hochberg", "hommel", "fdr", "BH", "BY",
    "none"),
  cl = NULL
)
}
\arguments{
\item{data}{A dataframe.\cr
Execute \code{data(Dataset_I)} and \code{View(Dataset_I)} for formats.}

\item{alpha_range}{A two-length numeric vector. Default is \code{c(0, 1)}. If the two values are the same, \eqn{\alpha} will be a fixed value.}

\item{lambda_range}{A two-length numeric vector. Default is \code{c(0, 10)}. If the two values are the same, \eqn{\lambda} will be a fixed value.}

\item{search_iterations}{A numeric scalar. Default is 500. Applied to random search for hyperparameter optimization.}

\item{random_seed}{A numeric scalar. Default is 123. Applied to random search for hyperparameter optimization.}

\item{continue}{Logical. Default is \code{FALSE}. Determines whether to implement CoCo when \code{QC_ST(data)} has indicated no significant batch effects across different batches.}

\item{sig_level.alpha}{A numeric scalar. Default is 0.05, which is unnecessary to be tuned for most of the time.}

\item{simul_method}{Default is \code{NULL}, which selects \code{fisher} when \eqn{\frac{n_{1}+n_{2}}{2} \ge 10} or else selects \code{HN}.\cr
\code{cauchy}, namely "Yu-Cauchy"; \code{fisher}, namely "Yu-Fisher".}

\item{p_value.adjust}{Default is \code{fdr}, which is equivalent to \code{BH} and is unnecessary to be tuned for most of the time.}

\item{cl}{Default is \code{NULL}, which uses all the CPU cores for parallel computing. Otherwise, it should be a numeric scalar.}
}
\value{
\item{hyperparameters}{A dataframe. The optimal hyperparameters.}
\item{VarFC_matrix}{A dataframe. The subject samples' variance fold changes before and after CoCo. Denoted as \eqn{\boldsymbol{V}}.}
\item{VarFC.all_mean}{A numeric scalar. The mean value of \code{VarFC_matrix}, namely \eqn{\mathrm{mean}(\boldsymbol{V})}, which determines whether the subject samples are overcorrected. We provide an acceptance range as reference: \eqn{(0.25,4)}.}
\item{corrected_data}{A dataframe. The format is the same as \code{data}.}
}
\description{
CoCo is used for inter-BEC, which employs the Graphical Elastic Net (GELNET) algorithm of the Gaussian Graphical Model (GGM).\cr
For the multivariate normal distribution \eqn{\mathrm{N}_{p}(\boldsymbol{\mu},\boldsymbol{\Sigma})}, the penalized Maximum Likelihood Estimator (MLE) of the precision matrix (inverse covariance matrix, namely \eqn{\boldsymbol{\Theta}=\boldsymbol{\Sigma}^{-1}}) by GELNET is:\cr
\deqn{\hat{\boldsymbol{\Theta}} = \arg\min_{\boldsymbol{\Theta}}\{-\mathrm{log}|\boldsymbol{\Theta}| + \mathrm{tr}(\boldsymbol{S}\boldsymbol{\Theta}) + \lambda(\alpha\|\boldsymbol{\Theta}-\boldsymbol{T}\|_{1} + \frac{1-\alpha}{2}\|\boldsymbol{\Theta}-\boldsymbol{T}\|^2_{2})\},}
where \eqn{\boldsymbol{S}} denotes the MLE of \eqn{\boldsymbol{\Sigma}};\cr
\eqn{\|.\|_{1}} denotes the matrix \eqn{L_{1}}-norm;\cr
\eqn{\|.\|_{2}} denotes the matrix \eqn{L_{2}}-norm (Frobenius norm);\cr
the hyperparameter \eqn{\alpha \in [0,1]};\cr
the hyperparameter \eqn{\lambda \in (0,+\infty)};\cr
\eqn{\boldsymbol{T}} denotes the target matrix, which is a positive semi-definite matrix. Here, we select \eqn{\boldsymbol{T}=\boldsymbol{I}_{p}} (see \strong{Note} for details).
}
\details{
Please read my preprint:
\itemize{
 \item{Guo, Z. High-dimensional Statistics Applications to Batch Effects in Metabolomics. \emph{arXiv} \strong{2024}. DOI: arXiv:2412.10196.}
}
}
\note{
\itemize{
 \item{GELNET is a sophisticated algorithm, which combines the Graphical Lasso (equivalent to \eqn{\alpha=1}) and the Graphical Ridge (equivalent to \eqn{\alpha=0}). When \eqn{\lambda \to 0}, \eqn{\hat{\boldsymbol{\Theta}} \to \boldsymbol{S}^{-1}}; When \eqn{\lambda \to +\infty}, \eqn{\hat{\boldsymbol{\Theta}} \to \boldsymbol{T}}.}
 \item{Several references have indicated that \eqn{\boldsymbol{T}=\boldsymbol{I}_{p}} performs well for most of the time. Additionally, \eqn{\boldsymbol{T}=\boldsymbol{I}_{p}} has an another advantage in CoCo (see the next note) comparing to other alternatives.}
 \item{If \code{QC_ST(data)} has indicated no significant batch effects across different batches, CoCo is no longer necessary. This is because redundant CoCo might be time-consuming, considering that the algorithm complexity is \eqn{O(Bp^3)}. Further, hyperparameter optimization of CoCo will obtain larger \eqn{\lambda}, which means that \code{data} and \code{CoCo(data)$corrected_data} will be almost the same, namely \eqn{\mathrm{mean}(\boldsymbol{V}) \approx 1}.}
}
}
\examples{
data(Dataset_I)
data <- Dataset_I
data.RF <- RF.correction(data)
data.RF.QC_ST <- QC_ST(data.RF)

RF_CoCo <- CoCo(data.RF)
data.RF_CoCo <- RF_CoCo$corrected_data
data.RF_CoCo.QC_ST <- QC_ST(data.RF_CoCo)
}
\references{
\itemize{
 \item{Kovács, S.; Ruckstuhl, T.; Obrist, H.; Bühlmann, P. Graphical Elastic Net and Target Matrices: Fast Algorithms and Software for Sparse Precision Matrix Estimation. \emph{arXiv} \strong{2021}. DOI: arXiv:2101.02148.}
 \item{Kheyri, A.; Bekker, A.; Arashi, M. High-Dimensional Precision Matrix Estimation through GSOS with Application in the Foreign Exchange Market. \emph{Mathematics} \strong{2022}, 10 (22). DOI: 10.3390/math10224232.}
 \item{Bekker, A.; Kheyri, A.; Arashi, M. A Computational Note on the Graphical Ridge in High-dimension. \emph{arXiv} \strong{2023}. DOI: arXiv:2312.15781.}
}
}
\seealso{
\code{\link{QC_ST}}, \code{GLassoElnetFast::\link[GLassoElnetFast]{gelnet}}.
}
\author{
Zhendong Guo (\email{guozhendong19@mails.ucas.ac.cn}).
}
