% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D_ratio.R
\name{compare_D_ratio}
\alias{compare_D_ratio}
\title{Dispersion-ratio (D-ratio) distribution of different data for comparison}
\usage{
compare_D_ratio(data1, ..., BEC_method, line_alpha = 0.8)
}
\arguments{
\item{data1}{A dataframe.\cr
Execute \code{data(Dataset_I)} and \code{View(Dataset_I)} for formats.}

\item{...}{Alternatives. More data (e.g., \code{data2}, \code{data3}) can be input.}

\item{BEC_method}{A character vector. The length must be the same as \code{c(data1, ...)}. For example, \code{c("raw", "RF", "XGBoost")}.}

\item{line_alpha}{A numeric scalar. Default is 0.8. Determines the curve line opacity in the cumulative distribution plot.}
}
\value{
A cumulative distribution plot.
}
\description{
The metric examines both QC samples and subject samples, whose formula is:\cr
\deqn{\mathrm{D\text{-}ratio}_{i}=\sqrt{\frac{\mathrm{Var}(y^{\mathrm{(QC)}}_{i})}{\mathrm{Var}(y^{\mathrm{(QC)}}_{i}) + \mathrm{Var}(y^{\mathrm{(ss)}}_{i})}},i=1,2,...,p,}\cr
where \eqn{y^{\mathrm{(QC)}}_{i}} denotes the intensity of QC samples' \eqn{i}th metabolite;\cr
\eqn{y^{\mathrm{(ss)}}_{i}} denotes the intensity of subject samples' \eqn{i}th metabolite;\cr
\eqn{p} denotes the variable (metabolite) number.
}
\note{
\itemize{
 \item{The acceptance criterion is 50\%.}
}
}
\examples{
data(Dataset_I)
data <- Dataset_I
data.RF <- RF.correction(data)
compare_D_ratio(data, data.RF, BEC_method = c("raw", "RF"))
}
\references{
\itemize{
 \item{Broadhurst, D.; Goodacre, R.; Reinke, S. N.; Kuligowski, J.; Wilson, I. D.; Lewis, M. R.; Dunn, W. B. Guidelines and considerations for the use of system suitability and quality control samples in mass spectrometry assays applied in untargeted clinical metabolomic studies. \emph{Metabolomics} \strong{2018}, 14 (6), 72. DOI: 10.1007/s11306-018-1367-3.}
}
}
\seealso{
\code{\link{D_ratio}}.
}
\author{
Zhendong Guo (\email{guozhendong19@mails.ucas.ac.cn}).
}
