% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter_plot.R
\name{compare_scatter_plot}
\alias{compare_scatter_plot}
\title{Scatter plot of different data for comparison}
\usage{
compare_scatter_plot(
  data1,
  ...,
  BEC_method,
  variable_index,
  sample_type = c("qc", "sample"),
  ylab = NULL,
  y.log10_trans = FALSE,
  y.lim = FALSE,
  y_min,
  y_max,
  point_size = 1.5,
  point_alpha = 0.8
)
}
\arguments{
\item{data1}{A dataframe.\cr
Execute \code{data(Dataset_I)} and \code{View(Dataset_I)} for formats.}

\item{...}{Alternatives. More data (e.g., \code{data2}, \code{data3}) can be input.}

\item{BEC_method}{A character vector. The length must be the same as \code{c(data1, ...)}. For example, \code{c("raw", "RF", "XGBoost")}.}

\item{variable_index}{A numeric scalar.}

\item{sample_type}{\code{qc} denotes QC samples; \code{sample} denotes subject samples.}

\item{ylab}{Default is \code{NULL}, which denotes the variable index. Otherwise, the variable name can be input.}

\item{y.log10_trans}{Logical. Default is \code{FALSE}.}

\item{y.lim}{Logical. Default is \code{FALSE}.}

\item{y_min}{A numeric scalar. Input only when \code{y.lim = TRUE}.}

\item{y_max}{A numeric scalar. Input only when \code{y.lim = TRUE}.}

\item{point_size}{A numeric scalar. Default is 1.5. Determines the point size in the scatter plot.}

\item{point_alpha}{A numeric scalar. Default is 0.8. Determines the point opacity in the scatter plot.}
}
\value{
A scatter plot.
}
\description{
A scatter plot for univariate visualization.
}
\examples{
data(Dataset_I)
data <- Dataset_I
data.RF <- RF.correction(data)
compare_scatter_plot(data, data.RF, BEC_method = c("raw", "RF"),
                     variable_index = 107, sample_type = "qc",
                     y.log10_trans = TRUE, y.lim = TRUE, y_min = 10^3, y_max = 10^5)
}
\seealso{
\code{\link{scatter_plot}}.
}
\author{
Zhendong Guo (\email{guozhendong19@mails.ucas.ac.cn}).
}
