% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D_ratio.R
\name{D_ratio}
\alias{D_ratio}
\title{D-ratio distribution of data}
\usage{
D_ratio(data, print_plot = TRUE, D_ratio_value = FALSE)
}
\arguments{
\item{data}{A dataframe. \strong{Use \code{data(Dataset_I)} for formats.}}

\item{print_plot}{Logical. Default is \code{TRUE}. Determines whether to plot the histogram and the cumulative distribution plot.}

\item{D_ratio_value}{Logical. Default is \code{FALSE}. Determines whether to output the D-ratio values.}
}
\value{
\item{`D-ratio: 0 ~ 50\%`}{A numeric scalar. The cumulative frequency of D-ratio within 50\%.}
\item{D_ratio}{A numeric vector. Output only when \code{D_ratio_value = TRUE}.}
\item{histogram}{Plotted only when \code{print_plot = TRUE}.}
\item{cumulative distribution plot}{Plotted only when \code{print_plot = TRUE}.}
}
\description{
Dispersion-ratio (D-ratio). The metric examines both QC samples and subject samples, whose formula is:\cr
\deqn{\mathrm{D\text{-}ratio}_{i}=\sqrt{\frac{\mathrm{Var}(y^{\mathrm{(QC)}}_{i})}{\mathrm{Var}(y^{\mathrm{(QC)}}_{i}) + \mathrm{Var}(y^{\mathrm{(ss)}}_{i})}},i=1,2,...,p,}\cr
where \eqn{y^{\mathrm{(QC)}}_{i}} denotes the intensity of QC samples' \eqn{i}th metabolite;\cr
\eqn{y^{\mathrm{(ss)}}_{i}} denotes the intensity of subject samples' \eqn{i}th metabolite;\cr
\eqn{p} denotes the variable (metabolite) number.
}
\note{
\itemize{
 \item{The acceptance criterion is 50\%.}
}
}
\examples{
data(Dataset_I)
data <- Dataset_I
D_ratio(data)
}
\references{
\itemize{
 \item{Broadhurst, D.; Goodacre, R.; Reinke, S. N.; Kuligowski, J.; Wilson, I. D.; Lewis, M. R.; Dunn, W. B. Guidelines and considerations for the use of system suitability and quality control samples in mass spectrometry assays applied in untargeted clinical metabolomic studies. \emph{Metabolomics} \strong{2018}, 14 (6), 72. DOI: 10.1007/s11306-018-1367-3.}
}
}
\seealso{
\code{\link{compare_D_ratio}}.
}
\author{
Zhendong Guo (\email{guozhendong19@mails.ucas.ac.cn}).
}
