% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_detection_by_PCA.R
\name{outlier_detection_by_PCA}
\alias{outlier_detection_by_PCA}
\title{Data-preprocessing: outlier detection by PCA for QC samples}
\usage{
outlier_detection_by_PCA(
  data,
  title = NULL,
  PCA_scale = TRUE,
  T2.distribution = c("Beta", "chisq", "T2.1", "T2.2"),
  print_plot = TRUE
)
}
\arguments{
\item{data}{A dataframe. \strong{Use \code{data(Dataset_I)} for formats.}}

\item{title}{Default is \code{NULL}. Otherwise, the batch name can be input.}

\item{PCA_scale}{Logical. Default is \code{TRUE}, which determines whether to autoscale (standardize) \code{data} and is unnecessary to be tuned for most of the time.}

\item{T2.distribution}{Default is \code{chisq}. Hotelling's \eqn{T^2} statistic follows four kinds of distribution in different fileds, whose strictness ranking (contrary to the size of the confidence ellipse) is:\cr
\code{Beta} > \code{chisq} > \code{T2.1} > \code{T2.2}.}

\item{print_plot}{Logical. Default is \code{TRUE}. Determines whether to plot the Hotelling's \eqn{T^2} chart and the SPE chart.}
}
\value{
\item{PC_num}{A numeric scalar. Determined by the Tracy-Widom distribution.}
\item{R2}{A numeric scalar. Cumulative \eqn{R^2} of the top \code{PC_num} principal components.}
\item{Hotelling_T2}{A list. Includes "95\%UCL ~ 99\%UCL" and "> 99\%UCL".}
\item{SPE}{A list. Includes "95\%UCL ~ 99\%UCL" and "> 99\%UCL".}
\item{intersect}{A list. The intersection of \code{Hotelling_T2} and \code{SPE}. Includes "95\%UCL ~ 99\%UCL" and "> 99\%UCL".}
\item{union}{A list. The union of \code{Hotelling_T2} and \code{SPE}. Includes "95\%UCL ~ 99\%UCL" and "> 99\%UCL".}
}
\description{
For outlier detection, multivariate statistical methods, such as Principal Component Analysis (PCA), are more powerful than univariate statistical methods due to the consideration for the correlations between variables.\cr
Here, we use the Tracy-Widom distribution to determine the principal component number of PCA.\cr
Then, we use the 95\% or 99\% upper control limit (UCL) of Hotelling's \eqn{T^2} statistic (also termed \eqn{D}-statistic) and Squared Prediction Errors (SPEs; also termed \eqn{Q}-statistic or DModX) to detect outliers.
}
\note{
\itemize{
 \item{The principle is based on the premise of the multivariate normal distribution. Hence, the function is suitable for QC samples instead of subject samples. Additionally, it should be used for QC samples batchwise instead of all the QC samples.}
 \item{It is an open problem to determine the principal component number of PCA. One of the feasible methods is the Tracy-Widom distribution, which is based on the Random Matrix Theory (RMT).}
 \item{Although Hotelling's \eqn{T^2} statistic strictly follows the Beta distribution, the Chi-squared distribution and the Hotelling's \eqn{T^2} distribution are still widely applied.}
 \item{By default, the samples whose \code{Hotelling_T2} and \code{SPE} are both larger than the 99\% UCL, namely \code{union}, are regarded as outliers.}
}
}
\examples{
data(Dataset_I.with_outliers)
data <- Dataset_I.with_outliers

if (length(data$order) != length(unique(data$order))){
  warning("The repeated values in 'order' should be corrected.")
}

# Obtain the batch names and the batch number.
batch.level <- unique(data$batch)
B <- length(batch.level)

# The variance of each variable should not be 0. Otherwise, eliminate them.
zero_var.index <- unique(unlist(lapply(1:B, function(b){
  QC_b <- subset(data, data$batch == batch.level[b] & data$type == 'qc')
  return(which(apply(QC_b[, -1:-4], 2, var) == 0))
})))

if (length(zero_var.index) > 1){
  data <- data[, -(4+zero_var.index)]
}

# Outlier detection for QC samples batchwise.
outlier_result <- lapply(1:B, function(b){
  QC_b <- subset(data, data$batch == batch.level[b] & data$type == 'qc')
  return(outlier_detection_by_PCA(QC_b, title = batch.level[b]))
})
names(outlier_result) <- batch.level

outliers <- unlist(sapply(1:B, function(b){
  which(data$order == outlier_result[[b]]$union$`> 99\%UCL`$order)
}))

# Obtain the data without outliers.
data <- data[-outliers, ]
}
\references{
\itemize{
 \item{Wikstrom, C.; Albano, C.; Eriksson, L.; Friden, H.; Johansson, E.; Nordahl, A.; Rannar, S.; Sandberg, M.; Kettaneh-Wold, N.; Wold, S. Multivariate process and quality monitoring applied to an electrolysis process Part I. Process supervision with multivariate control charts. \emph{Chemometrics and Intelligent Laboratory Systems} \strong{1998}, 42, 221-231.}
 \item{Saccenti, E.; Camacho, J. Determining the number of components in principal components analysis: A comparison of statistical, crossvalidation and approximated methods. \emph{Chemometrics and Intelligent Laboratory Systems} \strong{2015}, 149, 99-116. DOI: 10.1016/j.chemolab.2015.10.006.}
 \item{Blaise, B. J.; Correia, G. D. S.; Haggart, G. A.; Surowiec, I.; Sands, C.; Lewis, M. R.; Pearce, J. T. M.; Trygg, J.; Nicholson, J. K.; Holmes, E.; et al. Statistical analysis in metabolic phenotyping. \emph{Nature Protocols} \strong{2021}, 16 (9), 4299-4326. DOI: 10.1038/s41596-021-00579-1.}
}
}
\seealso{
\code{\link{PCA_plot}}, \code{stats::\link[stats]{prcomp}}.
}
\author{
Zhendong Guo (\email{guozhendong19@mails.ucas.ac.cn}).
}
