% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA_score_plot.R
\name{PCA_plot}
\alias{PCA_plot}
\title{Principal Component Analysis (PCA) score plot}
\usage{
PCA_plot(
  data,
  sample_type = c("all", "qc", "sample"),
  PCi = 1,
  PCj = 2,
  PCA_scale = TRUE,
  point_size = 1,
  point_alpha = 0.6
)
}
\arguments{
\item{data}{A dataframe.\cr
Execute \code{data(Dataset_I)} and \code{View(Dataset_I)} for formats.}

\item{sample_type}{\code{all} denotes all the samples; \code{qc} denotes QC samples; \code{sample} denotes subject samples.}

\item{PCi}{A numeric scalar. Default is 1. The \eqn{i}th principal component of PCA.}

\item{PCj}{A numeric scalar. Default is 2. The \eqn{j}th principal component of PCA.}

\item{PCA_scale}{Logical. Default is \code{TRUE}, which determines whether to autoscale (standardize) \code{data} and is unnecessary to be tuned for most of the time.}

\item{point_size}{A numeric scalar. Default is 1. Determines the point size in the scatter plot.}

\item{point_alpha}{A numeric scalar. Default is 0.6. Determines the point opacity in the scatter plot.}
}
\value{
A two-dimensional PCA score plot.
}
\description{
A two-dimensional PCA score plot for multivariate visualization.
}
\note{
\itemize{
 \item{PCA is a qualitative method for correction assessment, whose two-dimensional score plot might be limited and even misleading. Hence, pay attention to \eqn{R^2}.}
}
}
\examples{
data(Dataset_I)
data <- Dataset_I
PCA_plot(data, sample_type = "qc")
}
\seealso{
\code{\link{compare_PCA_plot}}, \code{stats::\link[stats]{prcomp}}.
}
\author{
Zhendong Guo (\email{guozhendong19@mails.ucas.ac.cn}).
}
