% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.RSD.R
\name{QC.RSD}
\alias{QC.RSD}
\title{Relative Standard Deviation (RSD) distribution}
\usage{
QC.RSD(data, print_plot = TRUE, RSD_value = FALSE)
}
\arguments{
\item{data}{A dataframe.\cr
Execute \code{data(Dataset_I)} and \code{View(Dataset_I)} for formats.}

\item{print_plot}{Logical. Default is \code{TRUE}. Determines whether to plot the histogram and the cumulative distribution plot.}

\item{RSD_value}{Logical. Default is \code{FALSE}. Determines whether to output the RSD values.}
}
\value{
\item{`QC.RSD: 0 ~ 30\%`}{A numeric scalar. The cumulative frequency of RSD within 30\%.}
\item{`QC.RSD: 0 ~ 20\%`}{A numeric scalar. The cumulative frequency of RSD within 20\%.}
\item{`QC.RSD: 0 ~ 15\%`}{A numeric scalar. The cumulative frequency of RSD within 15\%.}
\item{QC.RSD}{A numeric vector. Output only when \code{RSD_value = TRUE}.}
\item{histogram}{Plotted only when \code{print_plot = TRUE}.}
\item{cumulative distribution plot}{Plotted only when \code{print_plot = TRUE}.}
}
\description{
The metric only examines QC samples, whose formula is:\cr
\deqn{\mathrm{RSD}_{i}=\frac{\sqrt{\mathrm{Var}(y^{\mathrm{(QC)}}_{i})}}{\mathrm{mean}(y^{\mathrm{(QC)}}_{i})},i=1,2,...,p,}\cr
where \eqn{y^{\mathrm{(QC)}}_{i}} denotes the intensity of QC samples' \eqn{i}th metabolite;\cr
\eqn{p} denotes the variable (metabolite) number.
}
\note{
\itemize{
 \item{The acceptance criterion is 15\%, 20\%, and 30\%.}
}
}
\examples{
data(Dataset_I)
data <- Dataset_I
QC.RSD(data)
}
\references{
\itemize{
 \item{Broadhurst, D.; Goodacre, R.; Reinke, S. N.; Kuligowski, J.; Wilson, I. D.; Lewis, M. R.; Dunn, W. B. Guidelines and considerations for the use of system suitability and quality control samples in mass spectrometry assays applied in untargeted clinical metabolomic studies. \emph{Metabolomics} \strong{2018}, 14 (6), 72. DOI: 10.1007/s11306-018-1367-3.}
 \item{Dudzik, D.; Barbas-Bernardos, C.; Garcia, A.; Barbas, C. Quality assurance procedures for mass spectrometry untargeted metabolomics. a review. \emph{Journal of Pharmaceutical and Biomedical Analysis} \strong{2018}, 147, 149-173. DOI: 10.1016/j.jpba.2017.07.044.}
 \item{Han, W.; Li, L. Evaluating and minimizing batch effects in metabolomics. \emph{Mass Spectrometry Reviews} \strong{2020}, 41 (3), 421-442. DOI: 10.1002/mas.21672.}
 \item{Martens, A.; Holle, J.; Mollenhauer, B.; Wegner, A.; Kirwan, J.; Hiller, K. Instrumental Drift in Untargeted Metabolomics: Optimizing Data Quality with Intrastudy QC Samples. \emph{Metabolites} \strong{2023}, 13 (5). DOI: 10.3390/metabo13050665.}
}
}
\seealso{
\code{\link{compare_QC.RSD}}.
}
\author{
Zhendong Guo (\email{guozhendong19@mails.ucas.ac.cn}).
}
