% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_based_simultaneous_tests.R
\name{QC_ST}
\alias{QC_ST}
\title{QC-based simultaneous tests (QC-ST)}
\usage{
QC_ST(
  data,
  print_plot = TRUE,
  sig_level.alpha = 0.05,
  simul_method = c(NULL, "cauchy", "fisher", "HN"),
  p_value.adjust = c("bonferroni", "holm", "hochberg", "hommel", "fdr", "BH", "BY",
    "none")
)
}
\arguments{
\item{data}{A dataframe.\cr
Execute \code{data(Dataset_I)} and \code{View(Dataset_I)} for formats.}

\item{print_plot}{Logical. Default is \code{TRUE}. Determines whether to plot the heatmap and the undirected graph.}

\item{sig_level.alpha}{A numeric scalar. Default is 0.05, which is unnecessary to be tuned for most of the time.}

\item{simul_method}{Default is \code{NULL}, which selects \code{fisher} when \eqn{\mathrm{mean}\{n_{1},n_{2}\} \ge 10} or else selects \code{HN}.\cr
\code{cauchy}, namely "Yu-Cauchy"; \code{fisher}, namely "Yu-Fisher".}

\item{p_value.adjust}{Default is \code{fdr}, which is equivalent to \code{BH} and is unnecessary to be tuned for most of the time.}
}
\value{
\item{adjacency_matrix}{A matrix. 0 denotes no significant batch effects between the corresponding two batches. 1 denotes significant batch effects between the corresponding two batches.}
\item{`the pairs without significant batch effects`}{A character vector. Corresponds to the 0 elements of \code{adjacency_matrix}.}
\item{mean_test}{A dataframe. From the homogeneity tests of mean vectors.}
\item{cov_test}{A dataframe. From the homogeneity tests of covariance matrices.}
\item{simul_test}{A dataframe. From the simultaneous tests.}
\item{heatmap}{Plotted only when \code{print_plot = TRUE}. The red block means significant batch effects between the two batches; The blue block means no significant batch effects between the two batches.}
\item{undirected graph}{Plotted only when \code{print_plot = TRUE}. Existing no edge means significant batch effects between the two batches; Existing an edge means no significant batch effects between the two batches.}
}
\description{
The metric only examines QC samples.\cr
Considering the homogeneity of all the QC samples, we assume that if without batch effects, QC samples across different batches should follow the same multivariate normal distribution.\cr
QC-ST employs the simultaneous test of high-dimensional mean vectors and covariance matrices, which can be applied to the QC samples across different batches.\cr
The null hypothesis of each test is:\cr
\deqn{\boldsymbol{\mu}_{j_1}=\boldsymbol{\mu}_{j_2},\boldsymbol{\Sigma}_{j_1}=\boldsymbol{\Sigma}_{j_2},}
\deqn{1\le j_1<j_2\le B,}
where \eqn{\boldsymbol{\mu}} denotes the mean vector;\cr
\eqn{\boldsymbol{\Sigma}} denotes the covariance matrix;\cr
\eqn{B} denotes the batch number.
}
\details{
See our paper for details.
}
\note{
\itemize{
 \item{We recommend that the QC sample size of each batch should not be less than 10 to ensure high statistical powers.}
 \item{If \code{simul_test$q_value} \eqn{\ge \alpha_{\mathrm{sig}}}, take it as the standard, and ignore \code{mean_test} and \code{cov_test}.\cr
 Actually, the similar situation occurs in ANalysis Of Variance (ANOVA), where the \eqn{F}-test should be considered prior to the multiple tests.}
 \item{If \code{simul_test$q_value} \eqn{< \alpha_{\mathrm{sig}}}, \code{mean_test$q_value} and \code{cov_test$q_value} can be used to further determine which parameter does.}
 \item{If QC-ST suggests that batch effects between some two batches are significant after \code{\link{batch_ratio.correction}} (\code{method = ratio-A} or \code{mean}), the covariance matrices must have statistical significance.}
 \item{The principles of \code{mean_test} and \code{cov_test} corresponding to \code{cauchy} and \code{fisher} are the same, but slightly different from those corresponding to \code{HN}.}
}
}
\examples{
data(Dataset_I)
data <- Dataset_I
data.QC_ST <- QC_ST(data)
}
\references{
\itemize{
 \item{Chen, S. X.; Qin, Y.-L. A two-sample test for high-dimensional data with applications to gene-set testing. \emph{The Annals of Statistics} \strong{2010}, 38 (2). DOI: 10.1214/09-aos716.}
 \item{Li, J.; Chen, S. X. Two sample tests for high-dimensional covariance matrices. \emph{The Annals of Statistics} \strong{2012}, 40 (2). DOI: 10.1214/12-aos993.}
 \item{Hyodo, M.; Nishiyama, T. A simultaneous testing of the mean vector and the covariance matrix among two populations for high-dimensional data. \emph{Test} \strong{2017}, 27 (3), 680-699. DOI: 10.1007/s11749-017-0567-x.}
 \item{Miao, R.; Xu, K. Joint test for homogeneity of high-dimensional means and covariance matrices using maximum-type statistics. \emph{Communications in Statistics - Simulation and Computation} \strong{2022}, 53 (2), 972-992. DOI: 10.1080/03610918.2022.2037641.}
 \item{Yu, X.; Li, D.; Xue, L.; Li, R. Power-Enhanced Simultaneous Test of High-Dimensional Mean Vectors and Covariance Matrices with Application to Gene-Set Testing. \emph{Journal of the American Statistical Association} \strong{2022}, 118 (544), 2548-2561. DOI: 10.1080/01621459.2022.2061354.}
}
}
\seealso{
\code{PEtests::\link[PEtests]{meantest.cq}}, \code{PEtests::\link[PEtests]{covtest.lc}}, \code{PEtests::\link[PEtests]{simultest}}, \code{stats::\link[stats]{p.adjust}}.
}
\author{
Zhendong Guo (\email{guozhendong19@mails.ucas.ac.cn}).
}
