% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RF.correction.R
\name{RF.correction}
\alias{RF.correction}
\title{QC-based BEC algorithm: Random Forest (RF) regression}
\usage{
RF.correction(
  data,
  batch_ratio = c(NULL, "ratio-A", "median", "mean"),
  cor_variable_num = NULL,
  mtry = NULL,
  nodesize_percent = 0.2,
  ntree = 500,
  cl = NULL
)
}
\arguments{
\item{data}{A dataframe.\cr
Execute \code{data(Dataset_I)} and \code{View(Dataset_I)} for formats.}

\item{batch_ratio}{Default is \code{ratio-A} when the batch number \eqn{B>1}, or else is \code{NULL} when \eqn{B=1}.}

\item{cor_variable_num}{Default is \code{NULL}, which equals 10 when the variable number \eqn{p>10} or else equals \eqn{p-1}. Otherwise, it should be a numeric scalar. The hyperparameter usually has slight influence on correction.}

\item{mtry}{Default is \code{NULL}, which calls \code{randomForest::\link{tuneRF}} to automatically obtain a numeric vector for hyperparameter optimization. Otherwise, it should be a numeric scalar.}

\item{nodesize_percent}{A numeric scalar or vector. Default is 0.2.}

\item{ntree}{A numeric scalar. Default is 500. The hyperparameter usually has slight influence on correction.}

\item{cl}{Default is \code{NULL}, which uses all the CPU cores for parallel computing. Otherwise, it should be a numeric scalar.}
}
\value{
A dataframe. The corrected data, whose format is the same as \code{data}.
}
\description{
Briefly, the intra-BEC principle employs the regression model based on QC samples batchwise, namely:\cr
\deqn{y_{i}\sim f(x,\boldsymbol{y}_{(i)}\mid \boldsymbol{\theta}),i=1,2,…,p,}\cr
where \eqn{y_{i}} denotes the intensity of the \eqn{i}th metabolite;\cr
\eqn{x} denotes the injection order;\cr
\eqn{\boldsymbol{y}_{(i)}} denotes the several (i.e., \code{cor_variable_num}) variables with the highest correlations to the \eqn{i}th variable;\cr
\eqn{\boldsymbol{\theta}} denotes the optimal hyperparameters;\cr
\eqn{p} denotes the variable (metabolite) number.
}
\note{
\itemize{
 \item{Bootstrap is used for hyperparameter optimization instead of cross-validation.}
 \item{If \code{batch_ratio = NULL}, only intra-BEC will be implemented. If \code{batch_ratio = ratio-A} or \code{mean}, it can ensure the consistency among QC samples' mean vectors across different batches after correction.}
}
}
\examples{
data(Dataset_I)
data <- Dataset_I
data.RF <- RF.correction(data)
}
\references{
\itemize{
 \item{Luan, H.; Ji, F.; Chen, Y.; Cai, Z. statTarget: A streamlined tool for signal drift correction and interpretations of quantitative mass spectrometry-based omics data. \emph{Analytica Chimica Acta} \strong{2018}, 1036, 66-72. DOI: 10.1016/j.aca.2018.08.002.}
 \item{Fan, S.; Kind, T.; Cajka, T.; Hazen, S. L.; Tang, W. H. W.; Kaddurah-Daouk, R.; Irvin, M. R.; Arnett, D. K.; Barupal, D. K.; Fiehn, O. Systematic Error Removal Using Random Forest for Normalizing Large-Scale Untargeted Lipidomics Data. \emph{Analytical Chemistry} \strong{2019}, 91 (5), 3590-3596. DOI: 10.1021/acs.analchem.8b05592.}
}
}
\seealso{
\code{\link{batch_ratio.correction}}, \code{randomForest::\link[randomForest]{tuneRF}}, \code{randomForest::\link[randomForest]{randomForest}}.
}
\author{
Zhendong Guo (\email{guozhendong19@mails.ucas.ac.cn}).
}
