% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter_plot.R
\name{scatter_plot}
\alias{scatter_plot}
\title{Scatter plot}
\usage{
scatter_plot(
  data,
  variable_index,
  ylab = NULL,
  y.log10_trans = FALSE,
  y.lim = FALSE,
  y_min,
  y_max,
  point_size = 1.5,
  point_alpha = 0.8
)
}
\arguments{
\item{data}{A dataframe.\cr
Execute \code{data(Dataset_I)} and \code{View(Dataset_I)} for formats.}

\item{variable_index}{A numeric scalar.}

\item{ylab}{Default is \code{NULL}, which denotes the variable index. Otherwise, the variable name can be input.}

\item{y.log10_trans}{Logical. Default is \code{FALSE}.}

\item{y.lim}{Logical. Default is \code{FALSE}.}

\item{y_min}{A numeric scalar. Input only when \code{y.lim = TRUE}.}

\item{y_max}{A numeric scalar. Input only when \code{y.lim = TRUE}.}

\item{point_size}{A numeric scalar. Default is 1.5. Determines the point size in the scatter plot.}

\item{point_alpha}{A numeric scalar. Default is 0.8. Determines the point opacity in the scatter plot.}
}
\value{
A scatter plot.
}
\description{
A scatter plot for univariate visualization.
}
\examples{
data(Dataset_I)
data <- Dataset_I
scatter_plot(data, variable_index = 107, y.log10_trans = TRUE,
             y.lim = TRUE, y_min = 10^3, y_max = 10^5)
}
\seealso{
\code{\link{compare_scatter_plot}}.
}
\author{
Zhendong Guo (\email{guozhendong19@mails.ucas.ac.cn}).
}
