% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVR.correction.R
\name{SVR.correction}
\alias{SVR.correction}
\title{QC-based BEC algorithm: Support Vector Regression (SVR)}
\usage{
SVR.correction(
  data,
  batch_ratio = c(NULL, "ratio-A", "median", "mean"),
  cor_variable_num = NULL,
  gamma = NULL,
  cost = 1,
  epsilon = 0.1,
  cl = NULL
)
}
\arguments{
\item{data}{A dataframe. \strong{Use \code{data(Dataset_I)} for formats.}}

\item{batch_ratio}{Default is \code{ratio-A} when the batch number \eqn{B>1}, otherwise default is \code{NULL} when \eqn{B=1}. Noted that if \code{batch_ratio = NULL}, only intra-BEC will be implemented.}

\item{cor_variable_num}{Default is \code{NULL}, which equals to 10 when the variable number \eqn{p>10} or else equals to \eqn{p-1}. Otherwise, it should be a numeric scalar. The hyperparameter usually has slight influence on correction. Denoted as \eqn{p'}.}

\item{gamma}{Default is \code{NULL}, which is \eqn{\frac{1}{p'+1}·} \code{c(1,2,4)} for hyperparameter optimization. Otherwise, it should be a numeric scalar or vector.}

\item{cost}{A numeric scalar or vector. Default is 1.}

\item{epsilon}{A numeric scalar or vector. Default is 0.1. The hyperparameter usually has slight influence on correction.}

\item{cl}{Default is \code{NULL}, which uses all the CPU cores for parallel computing. Otherwise, it should be a numeric scalar.}
}
\value{
A dataframe. The corrected data, whose format is the same as \code{data}.
}
\description{
Briefly, the intra-BEC principle employs the regression model based on QC samples batchwise, namely:\cr
\deqn{y_{i}\sim f(x,\boldsymbol{y}_{(i)}\mid \boldsymbol{\theta}),i=1,2,…,p,}\cr
where \eqn{y_{i}} denotes the intensity of the \eqn{i}th metabolite;\cr
\eqn{x} denotes the injection order of the corresponding batch;\cr
\eqn{\boldsymbol{y}_{(i)}} denotes the several (i.e., \code{cor_variable_num}) variables with the highest correlations to \eqn{y_{i}};\cr
\eqn{\boldsymbol{\theta}} denotes the optimal hyperparameters;\cr
\eqn{p} denotes the variable (metabolite) number.
}
\note{
\itemize{
 \item{5-fold-cross-validation (CV) is used for hyperparameter optimization.}
 \item{When \code{batch_ratio = ratio-A} or \code{mean}, batch-ratio can ensure the consistency among QC samples' mean vectors across different batches after correction.}
}
}
\examples{
data(Dataset_I)
data <- Dataset_I
data.SVR <- SVR.correction(data)
}
\references{
\itemize{
 \item{Kuligowski, J.; Sanchez-Illana, A.; Sanjuan-Herraez, D.; Vento, M.; Quintas, G. Intra-batch effect correction in liquid chromatography-mass spectrometry using quality control samples and support vector regression (QC-SVRC). \emph{Analyst} \strong{2015}, 140 (22), 7810-7817. DOI: 10.1039/c5an01638j.}
 \item{Shen, X.; Gong, X.; Cai, Y.; Guo, Y.; Tu, J.; Li, H.; Zhang, T.; Wang, J.; Xue, F.; Zhu, Z.-J. Normalization and integration of large-scale metabolomics data using support vector regression. \emph{Metabolomics} \strong{2016}, 12 (5). DOI: 10.1007/s11306-016-1026-5.}
 \item{Ding, X.; Yang, F.; Chen, Y.; Xu, J.; He, J.; Zhang, R.; Abliz, Z. Norm ISWSVR: A Data Integration and Normalization Approach for Large-Scale Metabolomics. \emph{Anal Chem} \strong{2022}, 94 (21), 7500-7509. DOI: 10.1021/acs.analchem.1c05502.}
}
}
\seealso{
\code{\link{batch_ratio.correction}}, \code{e1071::\link[e1071]{tune}}, \code{e1071::\link[e1071]{svm}}.
}
\author{
Zhendong Guo (\email{guozhendong19@mails.ucas.ac.cn}).
}
