% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TIGER.correction.R
\name{TIGER.correction}
\alias{TIGER.correction}
\title{QC-based BEC algorithm: Technical variation elImination with ensemble learninG architEctuRe (TIGER)}
\usage{
TIGER.correction(
  data,
  batch_ratio = c(NULL, "ratio-A", "median", "mean"),
  cor_variable_num = NULL,
  mtry_percent = seq(0.2, 0.8, 0.2),
  nodesize_percent = 0.2,
  ntree = 500,
  cl = NULL
)
}
\arguments{
\item{data}{A dataframe. \strong{Use \code{data(Dataset_I)} for formats.}}

\item{batch_ratio}{Default is \code{ratio-A} when the batch number \eqn{B>1}, otherwise default is \code{NULL} when \eqn{B=1}. Noted that if \code{batch_ratio = NULL}, only intra-BEC will be implemented.}

\item{cor_variable_num}{Default is \code{NULL}, which equals to 10 when the variable number \eqn{p>10} or else equals to \eqn{p-1}. Otherwise, it should be a numeric scalar. The hyperparameter usually has slight influence on correction.}

\item{mtry_percent}{A numeric scalar or vector. Default is \code{seq(0.2, 0.8, 0.2)}.}

\item{nodesize_percent}{A numeric scalar or vector. Default is 0.2.}

\item{ntree}{A numeric scalar or vector. Default is 500. The hyperparameter usually has slight influence on correction.}

\item{cl}{Default is \code{NULL}, which uses all the CPU cores for parallel computing. Otherwise, it should be a numeric scalar.}
}
\value{
A dataframe. The corrected data, whose format is the same as \code{data}.
}
\description{
Briefly, the intra-BEC principle is divided into two parts based on QC samples batchwise: the base model and the meta model. The former employs Random Forest (RF) to establish the regression model similar to \code{\link{SVR.correction}}, \code{\link{RF.correction}}, and \code{\link{XGBoost.correction}}, and the latter integrates several base models with different hyperparameters weighted by the loss function.
}
\note{
\itemize{
 \item{TIGER has a more complex correction principle than \code{\link{SVR.correction}}, \code{\link{RF.correction}}, and \code{\link{XGBoost.correction}} (see the reference below for details).}
 \item{5-fold-cross-validation (CV) is used for hyperparameter optimization.}
 \item{When \code{batch_ratio = ratio-A} or \code{mean}, batch-ratio can ensure the consistency among QC samples' mean vectors across different batches after correction.}
}
}
\examples{
data(Dataset_I)
data <- Dataset_I
data.TIGER <- TIGER.correction(data)
}
\references{
\itemize{
 \item{Han, S.; Huang, J.; Foppiano, F.; Prehn, C.; Adamski, J.; Suhre, K.; Li, Y.; Matullo, G.; Schliess, F.; Gieger, C.; et al. TIGER: technical variation elimination for metabolomics data using ensemble learning architecture. \emph{Briefings in Bioinformatics} \strong{2022}, 23 (2). DOI: 10.1093/bib/bbab535.}
}
}
\seealso{
\code{\link{batch_ratio.correction}}, \code{TIGERr::\link[TIGERr]{run_TIGER}}, \code{randomForest::\link[randomForest]{randomForest}}.
}
\author{
Zhendong Guo (\email{guozhendong19@mails.ucas.ac.cn}).
}
