% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XGBoost.correction.R
\name{XGBoost.correction}
\alias{XGBoost.correction}
\title{QC-based BEC algorithm: eXtreme Gradient Boost (XGBoost) regression}
\usage{
XGBoost.correction(
  data,
  batch_ratio = c(NULL, "ratio-A", "median", "mean"),
  cor_variable_num = NULL,
  nrounds = 50,
  eta = c(0.05, 0.1),
  gamma = 1,
  min_child_weight = 2,
  max_depth = 4,
  subsample = 0.7,
  colsample_bytree = 0.8,
  cl = NULL
)
}
\arguments{
\item{data}{A dataframe. \strong{Use \code{data(Dataset_I)} for formats.}}

\item{batch_ratio}{Default is \code{ratio-A} when the batch number \eqn{B>1}, otherwise default is \code{NULL} when \eqn{B=1}. Noted that if \code{batch_ratio = NULL}, only intra-BEC will be implemented.}

\item{cor_variable_num}{Default is \code{NULL}, which equals to 10 when the variable number \eqn{p>10} or else equals to \eqn{p-1}. Otherwise, it should be a numeric scalar. The hyperparameter usually has slight influence on correction.}

\item{nrounds}{A numeric scalar or vector. Default is 50.}

\item{eta}{A numeric scalar or vector. Default is \code{c(0.05, 0.1)}.}

\item{gamma}{A numeric scalar or vector. Default is 1. The hyperparameter usually has slight influence on correction.}

\item{min_child_weight}{A numeric scalar or vector. Default is 2.}

\item{max_depth}{A numeric scalar or vector. Default is 4. The hyperparameter is similar to \code{nodesize} or \code{maxnodes} in \code{randomForest::\link{randomForest}}.}

\item{subsample}{A numeric scalar or vector. Default is 0.7. The hyperparameter is applied without replacement, contrary to bagging (\code{replace = TRUE}) in \code{randomForest::\link{randomForest}}.}

\item{colsample_bytree}{A numeric scalar or vector. Default is 0.8. The hyperparameter is similar to \code{mtry} in \code{randomForest::\link{randomForest}}.}

\item{cl}{Default is \code{NULL}, which uses all the CPU cores for parallel computing. Otherwise, it should be a numeric scalar.}
}
\value{
A dataframe. The corrected data, whose format is the same as \code{data}.
}
\description{
Briefly, the intra-BEC principle employs the regression model based on QC samples batchwise, namely:\cr
\deqn{y_{i}\sim f(x,\boldsymbol{y}_{(i)}\mid \boldsymbol{\theta}),i=1,2,…,p,}\cr
where \eqn{y_{i}} denotes the intensity of the \eqn{i}th metabolite;\cr
\eqn{x} denotes the injection order of the corresponding batch;\cr
\eqn{\boldsymbol{y}_{(i)}} denotes the several (i.e., \code{cor_variable_num}) variables with the highest correlations to \eqn{y_{i}};\cr
\eqn{\boldsymbol{\theta}} denotes the optimal hyperparameters;\cr
\eqn{p} denotes the variable (metabolite) number.
}
\note{
\itemize{
 \item{XGBoost has more hyperparameters when establishing the regression model, such that it is more sophisticated and powerful to handle different scales of datasets.}
 \item{5-fold-cross-validation (CV) is used for hyperparameter optimization.}
 \item{When \code{batch_ratio = ratio-A} or \code{mean}, batch-ratio can ensure the consistency among QC samples' mean vectors across different batches after correction.}
}
}
\examples{
data(Dataset_I)
data <- Dataset_I
data.XGBoost <- XGBoost.correction(data)
}
\references{
\itemize{
 \item{Chen, T.; Guestrin, C.; Machinery, A. C. XGBoost: A Scalable Tree Boosting System. In \emph{KDD'16: Proceedings of the 22Nd ACM SIGKDD International Conference on Knowledge Discovery and Data Mining}, 2016; pp 785-794. DOI: 10.1145/2939672.2939785.}
 \item{Bentéjac, C.; Csrg, A.; Martínez-Muoz, G. A Comparative Analysis of XGBoost. \emph{Arxiv} \strong{2019}. DOI: arXiv:1911.01914.}
}
}
\seealso{
\code{\link{batch_ratio.correction}}, \code{caret::\link[caret]{train}}, \code{xgboost::xgboost}.
}
\author{
Zhendong Guo (\email{guozhendong19@mails.ucas.ac.cn}).
}
